+++ src/commands/index.js
@@ -1,0 +1,3 @@
+export * from './help.js';
+export * from './nwd.js';
+export * from './os.js';
